LibMedia+LibGfx: Remove owned storage from VideoFrame
This was unused except in tests after the previous commit.
